Add custom Tcl_ObjType (pgStatement) that caches StatementData in its intrep.
Transport prepared statements with the ConnectionData when it's detached and attached.

Add a statements hash to the ConnectionData and a linked list of referencing Tcl_Objs to the StatementData to make the refcounting work for the freezing and thawing of prepared statements.

FossilOrigin-Name: 17ad0ca0d73db4e6459cf0924ad94da62ea28fd2a0d95c6532f4af0a7f5bd17a